Add the function to replay to IFS ICs; update setup scripts to python 3; some other changes related to model changes to model code change

+++ ush/rocoto/setup_expt_replay_shield.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 ###############################################################
 # < next few lines under version control, D O  N O T  E D I T >
@@ -28,9 +28,9 @@ def makedirs_if_missing(d):
 def create_EXPDIR():
 
     makedirs_if_missing(expdir)
-    configs = glob.glob('%s/config.*' % configdir)
+    configs = glob.glob(f'{configdir}/config.*')
     if len(configs) == 0:
-        msg = 'no config files found in %s' % configdir
+        msg = f'no config files found in {configdir}'
         raise IOError(msg)
     for config in configs:
         shutil.copy(config, expdir)
@@ -47,24 +47,24 @@ def create_COMROT():
 
 def edit_baseconfig():
 
-    base_config = '%s/config.base' % expdir
+    base_config = f'{expdir}/config.base'
 
     here = os.path.dirname(__file__)
     top = os.path.abspath(os.path.join(os.path.abspath(here), '../..'))
 
-    # make a copy of the default before editing
-    shutil.copy(base_config, base_config + '.default')
+    if os.path.exists(base_config):
+        os.unlink(base_config)
 
-    print '\nSDATE = %s\nEDATE = %s' % (idate, edate)
-    with open(base_config + '.default', 'rt') as fi:
-        with open(base_config + '.new', 'wt') as fo:
+    print(f'\nSDATE = {idate}\nEDATE = {edate}')
+    with open(base_config + '.gfdl', 'rt') as fi:
+        with open(base_config, 'wt') as fo:
             for line in fi:
                 line = line.replace('@MACHINE@', machine.upper()) \
                     .replace('@PSLOT@', pslot) \
                     .replace('@SDATE@', idate.strftime('%Y%m%d%H')) \
                     .replace('@FDATE@', fdate.strftime('%Y%m%d%H')) \
                     .replace('@EDATE@', edate.strftime('%Y%m%d%H')) \
-                    .replace('@CASECTL@', 'C%d' % res) \
+                    .replace('@CASECTL@', f'C{res}') \
                     .replace('@HOMEgfs@', top) \
                     .replace('@BASE_GIT@', base_git) \
                     .replace('@DMPDIR@', dmpdir) \
@@ -80,29 +80,28 @@ def edit_baseconfig():
                     .replace('@PARTITION_BATCH@', partition_batch) \
                     .replace('@EXP_WARM_START@', exp_warm_start) \
                     .replace('@ICDUMP@', icdump) \
+                    .replace('@ICSTYP@', icstyp) \
                     .replace('@MODE@', 'replay') \
                     .replace('@CHGRP_RSTPROD@', chgrp_rstprod) \
                     .replace('@CHGRP_CMD@', chgrp_cmd) \
                     .replace('@HPSSARCH@', hpssarch) \
-                    .replace('@gfs_cyc@', '%d' % gfs_cyc) \
-                    .replace('@gfs_delay@', '%d' % gfs_delay)
+                    .replace('@gfs_cyc@', f'{gfs_cyc}') \
+                    .replace('@gfs_delay@', f'{gfs_delay}')
                 if expdir is not None:
                     line = line.replace('@EXPDIR@', os.path.dirname(expdir))
                 if comrot is not None:
                     line = line.replace('@ROTDIR@', os.path.dirname(comrot))
-                if icsdir is not None: 
+                if icsdir is not None:
                     line = line.replace('@ICSDIR@', icsdir)
                 else:
                     line = line.replace('@ICSDIR@', os.path.join(comrot, 'ICS'))
                 fo.write(line)
-    os.unlink(base_config)
-    os.rename(base_config + '.new', base_config)
 
-    print ''
-    print 'EDITED:  %s/config.base as per user input.' % expdir
-    print 'DEFAULT: %s/config.base.default is for reference only.' % expdir
-    print 'Please verify and delete the default file before proceeding.'
-    print ''
+    print('')
+    print(f'EDITED:  {expdir}/config.base as per user input.')
+    print(f'DEFAULT: {expdir}/config.base.gfdl is for reference only.')
+    print('Please verify and delete the default file before proceeding.')
+    print('')
 
     return
 
@@ -111,7 +110,8 @@ if __name__ == '__main__':
 
     description = '''Setup files and directories to start a GFS parallel.
 Create EXPDIR, copy config files
-Create COMROT experiment directory structure'''
+Create COMROT experiment directory structure,
+link initial condition files from $ICSDIR to $COMROT'''
 
     parser = ArgumentParser(description=description, formatter_class=ArgumentDefaultsHelpFormatter)
     parser.add_argument('--pslot', help='parallel experiment name', type=str, required=False, default='test')
@@ -119,6 +119,7 @@ Create COMROT experiment directory structure'''
     parser.add_argument('--comrot', help='full path to COMROT', type=str, required=False, default=None)
     parser.add_argument('--expdir', help='full path to EXPDIR', type=str, required=False, default=None)
     parser.add_argument('--icsdir', help='full path to ICSDIR', type=str, required=False, default=None)
+    parser.add_argument('--icstyp', help='external IC type', type=str, required=False, default='gfs')
     parser.add_argument('--icdump', help='initial condition dump', type=str, choices=['gdas', 'gfs'], default='gdas', required=False)
     parser.add_argument('--idate', help='starting date of experiment, initial conditions must exist!', type=str, required=True)
     parser.add_argument('--edate', help='end date experiment', type=str, required=True)
@@ -144,13 +145,13 @@ Create COMROT experiment directory structure'''
     gfs_delay = args.gfs_delay
 
     res = args.res
-    icsdir = args.icsdir if args.icsdir is None else os.path.join(args.comrot, 'ICS')
+    icsdir = args.icsdir if args.icsdir is not None else os.path.join(args.comrot, 'ICS')
     comrot = args.comrot if args.comrot is None else os.path.join(args.comrot, pslot)
     expdir = args.expdir if args.expdir is None else os.path.join(args.expdir, pslot)
     icdump = args.icdump
+    icstyp = args.icstyp
     partition = args.partition
     start = args.start
-    hpssarch = args.hpssarch
 
     # Set restart setting in config.base
     if start == 'cold':
@@ -181,7 +182,7 @@ Create COMROT experiment directory structure'''
             queue_service = 'dev2_transfer'
         chgrp_rstprod = 'YES'
         chgrp_cmd = 'chgrp rstprod'
-        #hpssarch = 'YES'
+        hpssarch = 'YES'
     elif machine == 'WCOSS_C':
         base_git = '/gpfs/hps3/emc/global/noscrub/emc.glopara/git'
         base_svn = '/gpfs/hps3/emc/global/noscrub/emc.glopara/svn'
@@ -198,13 +199,13 @@ Create COMROT experiment directory structure'''
         partition_batch = ''
         chgrp_rstprod = 'YES'
         chgrp_cmd = 'chgrp rstprod'
-        #hpssarch = 'YES'
+        hpssarch = 'YES'
     elif machine == 'HERA':
         base_git = '/scratch1/NCEPDEV/global/glopara/git'
         base_svn = '/scratch1/NCEPDEV/global/glopara/svn'
         dmpdir = '/scratch1/NCEPDEV/global/glopara/dump'
         nwprod = '/scratch1/NCEPDEV/global/glopara/nwpara'
-        comroot = '/scratch1/NCEPDEV/global/glopara/com'
+        comroot = '/scratch1/NCEPDEV/rstprod/com'
         homedir = '/scratch2/GFDL/gfdlscr/$USER'
         stmp = '/scratch2/NCEPDEV/stmp1/$USER'
         ptmp = '/scratch1/NCEPDEV/stmp4/$USER'
@@ -215,11 +216,11 @@ Create COMROT experiment directory structure'''
         partition_batch = ''
         chgrp_rstprod = 'YES'
         chgrp_cmd = 'chgrp rstprod'
-        #hpssarch = 'YES'
+        hpssarch = 'YES'
     elif machine == 'ORION':
         base_git = '/work/noaa/global/glopara/git'
         base_svn = '/work/noaa/global/glopara/svn'
-        dmpdir = '/work/noaa/global/glopara/dump'
+        dmpdir = '/work/noaa/rstprod/dump'
         nwprod = '/work/noaa/global/glopara/nwpara'
         comroot = '/work/noaa/global/glopara/com'
         homedir = '/work/noaa/global/$USER'
@@ -230,20 +231,27 @@ Create COMROT experiment directory structure'''
         queue = 'batch'
         queue_service = 'service'
         partition_batch = 'orion'
-        chgrp_rstprod = 'NO'          # No rstprod on Orion
-        chgrp_cmd = 'ls'
+        chgrp_rstprod = 'YES'
+        chgrp_cmd = 'chgrp rstprod'
         hpssarch = 'NO'
 
+    if args.icsdir is not None and not os.path.exists(icsdir):
+        msg = f'Initial conditions do not exist in {icsdir}'
+        raise IOError(msg)
+
     # COMROT directory
-    create_comrot = True
-    if os.path.exists(comrot):
-        print
-        print 'COMROT already exists in %s' % comrot
-        print
-        overwrite_comrot = raw_input('Do you wish to over-write COMROT [y/N]: ')
-        create_comrot = True if overwrite_comrot in ['y', 'yes', 'Y', 'YES'] else False
-        if create_comrot:
-            shutil.rmtree(comrot)
+    if args.icsdir is None:
+       create_comrot = False
+    else:
+       create_comrot = True
+       if os.path.exists(comrot):
+           print()
+           print(f'COMROT already exists in {comrot}')
+           print()
+           overwrite_comrot = input('Do you wish to over-write COMROT [y/N]: ')
+           create_comrot = True if overwrite_comrot in ['y', 'yes', 'Y', 'YES'] else False
+           if create_comrot:
+              shutil.rmtree(comrot)
 
     if create_comrot:
         create_COMROT()
@@ -251,10 +259,10 @@ Create COMROT experiment directory structure'''
     # EXP directory
     create_expdir = True
     if os.path.exists(expdir):
-        print
-        print 'EXPDIR already exists in %s' % expdir
-        print
-        overwrite_expdir = raw_input('Do you wish to over-write EXPDIR [y/N]: ')
+        print()
+        print(f'EXPDIR already exists in {expdir}')
+        print()
+        overwrite_expdir = input('Do you wish to over-write EXPDIR [y/N]: ')
         create_expdir = True if overwrite_expdir in ['y', 'yes', 'Y', 'YES'] else False
         if create_expdir:
             shutil.rmtree(expdir)
